Convert version to double for consistency with JS library

+++ src/main/java/org/evtstore/Aggregate.java
@@ -2,5 +2,5 @@ package org.evtstore;
 
 public class Aggregate extends Object {
   public String aggregateId = "";
-  public Integer version = 0;
+  public Double version = 0.0;
 }
